requirements.txt
streamlit
opencv-python-headless
numpy
scipy
Pillow